Extract authorize_as_authenticated into a shared trait
Moves the duplicated permission callback used by both
NotificationPreferencesRestController and PushTokenRestController into a
new AuthorizesPushNotificationRequests trait, mirroring the existing
ConvertsExceptionsToWpError trait pattern. PushNotifications is now
looked up via the container in one place rather than re-implemented per
controller, and the now-unused dependency injection of PushNotifications
in NotificationPreferencesRestController has been dropped.

Follow-up to https://github.com/woocommerce/woocommerce/pull/64351#discussion_r3218483918.
