+++ plugins/woocommerce/src/Internal/PushNotifications/Controllers/NotificationPreferencesRestController.php
@@ -6,8 +6,8 @@ namespace Automattic\WooCommerce\Internal\PushNotifications\Controllers;
 
 defined( 'ABSPATH' ) || exit;
 
-use Automattic\WooCommerce\Internal\PushNotifications\PushNotifications;
 use Automattic\WooCommerce\Internal\PushNotifications\Services\NotificationPreferencesService;
+use Automattic\WooCommerce\Internal\PushNotifications\Traits\AuthorizesPushNotificationRequests;
 use Automattic\WooCommerce\Internal\PushNotifications\Traits\ConvertsExceptionsToWpError;
 use Automattic\WooCommerce\Internal\RestApiControllerBase;
 use Exception;
@@ -24,6 +24,7 @@ use WP_REST_Server;
  * @since 10.8.0
  */
 class NotificationPreferencesRestController extends RestApiControllerBase {
+	use AuthorizesPushNotificationRequests;
 	use ConvertsExceptionsToWpError;
 
 	/**
@@ -48,28 +49,16 @@ class NotificationPreferencesRestController extends RestApiControllerBase {
 	private NotificationPreferencesService $preferences_service;
 
 	/**
-	 * The push notifications module enablement gate.
-	 *
-	 * @var PushNotifications
-	 */
-	private PushNotifications $push_notifications;
-
-	/**
 	 * Initialize injected dependencies.
 	 *
 	 * @internal
 	 *
 	 * @param NotificationPreferencesService $preferences_service The preferences service.
-	 * @param PushNotifications              $push_notifications  The push notifications module.
 	 *
 	 * @since 10.8.0
 	 */
-	final public function init(
-		NotificationPreferencesService $preferences_service,
-		PushNotifications $push_notifications
-	): void {
+	final public function init( NotificationPreferencesService $preferences_service ): void {
 		$this->preferences_service = $preferences_service;
-		$this->push_notifications  = $push_notifications;
 	}
 
 	/**
@@ -152,37 +141,6 @@ class NotificationPreferencesRestController extends RestApiControllerBase {
 		}
 
 		return new WP_REST_Response( $merged, WP_Http::OK );
-	}
-
-	/**
-	 * Checks user is authenticated and authorized to access this endpoint.
-	 *
-	 * @since 10.8.0
-	 *
-	 * @param WP_REST_Request $request The request object.
-	 * @phpstan-param WP_REST_Request<array<string, mixed>> $request
-	 * @return bool|WP_Error
-	 */
-	public function authorize_as_authenticated( WP_REST_Request $request ) {
-		if ( ! get_current_user_id() ) {
-			return new WP_Error(
-				'woocommerce_rest_cannot_view',
-				__( 'Sorry, you are not allowed to do that.', 'woocommerce' ),
-				array( 'status' => rest_authorization_required_code() )
-			);
-		}
-
-		if ( ! $this->push_notifications->should_be_enabled() ) {
-			return false;
-		}
-
-		$has_valid_role = array_reduce(
-			PushNotifications::ROLES_WITH_PUSH_NOTIFICATIONS_ENABLED,
-			fn ( $carry, $role ) => $this->check_permission( $request, $role ) === true ? true : $carry,
-			false
-		);
-
-		return $has_valid_role ? true : false;
 	}
 
 	/**
